requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.25
psycopg==3.3.2
pydantic==2.10.4
dbt-core==1.8.7
dbt-postgres==1.8.2
pytest==8.3.4
telethon==1.36.0
python-dotenv==1.0.1
ultralytics==8.3.0
opencv-python==4.10.0.84
pandas
dagster==1.8.13
dagster-webserver==1.8.13
dagster-dbt==0.23.1
